requirements.txt
torchvision~=0.10.0
torch~=1.9.0
Pillow~=8.2.0
matplotlib~=3.4.2
tqdm~=4.61.1